DeepLPF/requirements.txt
numpy==1.19.2
scipy==1.5.2
torch==1.7.1
scikit_image==0.18.1
torchvision==0.8.2
matplotlib==3.3.2
imageio==2.9.0
Pillow==8.1.0
skimage==0.0
